requirements.txt
streamlit>=1.35.0
huggingface_hub>=0.23.0
pymupdf>=1.24.0
python-docx>=1.1.0
python-dotenv>=1.0.0